+++ net/minecraft/block/WitherRoseBlock.java
@@ -45,7 +45,7 @@ extends FlowerBlock {
         double e = (double)pos.getZ() + vec3d.z;
         for (int i = 0; i < 3; ++i) {
             if (!random.nextBoolean()) continue;
-            world.addParticle(ParticleTypes.SMOKE, d + (double)(random.nextFloat() / 5.0f), (double)pos.getY() + (0.5 - (double)random.nextFloat()), e + (double)(random.nextFloat() / 5.0f), 0.0, 0.0, 0.0);
+            world.addParticle(ParticleTypes.SMOKE, d + random.nextDouble() / 5.0, (double)pos.getY() + (0.5 - random.nextDouble()), e + random.nextDouble() / 5.0, 0.0, 0.0, 0.0);
         }
     }
