+++ net/minecraft/block/entity/EnchantingTableBlockEntity.java
@@ -59,7 +59,7 @@ Tickable {
         float g;
         this.pageTurningSpeed = this.nextPageTurningSpeed;
         this.field_11963 = this.field_11964;
-        PlayerEntity playerEntity = this.world.getClosestPlayer((double)((float)this.pos.getX() + 0.5f), (double)((float)this.pos.getY() + 0.5f), (double)((float)this.pos.getZ() + 0.5f), 3.0, false);
+        PlayerEntity playerEntity = this.world.getClosestPlayer((double)this.pos.getX() + 0.5, (double)this.pos.getY() + 0.5, (double)this.pos.getZ() + 0.5, 3.0, false);
         if (playerEntity != null) {
             double d = playerEntity.getX() - ((double)this.pos.getX() + 0.5);
             double e = playerEntity.getZ() - ((double)this.pos.getZ() + 0.5);
