+++ net/minecraft/entity/ai/goal/DoorInteractGoal.java
@@ -78,15 +78,15 @@ extends Goal {
     @Override
     public void start() {
         this.shouldStop = false;
-        this.xOffset = (float)((double)((float)this.doorPos.getX() + 0.5f) - this.mob.getX());
-        this.zOffset = (float)((double)((float)this.doorPos.getZ() + 0.5f) - this.mob.getZ());
+        this.xOffset = (float)((double)this.doorPos.getX() + 0.5 - this.mob.getX());
+        this.zOffset = (float)((double)this.doorPos.getZ() + 0.5 - this.mob.getZ());
     }
 
     @Override
     public void tick() {
         float g;
-        float f = (float)((double)((float)this.doorPos.getX() + 0.5f) - this.mob.getX());
-        float h = this.xOffset * f + this.zOffset * (g = (float)((double)((float)this.doorPos.getZ() + 0.5f) - this.mob.getZ()));
+        float f = (float)((double)this.doorPos.getX() + 0.5 - this.mob.getX());
+        float h = this.xOffset * f + this.zOffset * (g = (float)((double)this.doorPos.getZ() + 0.5 - this.mob.getZ()));
         if (h < 0.0f) {
             this.shouldStop = true;
         }
