+++ net/minecraft/fluid/LavaFluid.java
@@ -54,9 +54,9 @@ extends FlowableFluid {
         BlockPos blockPos = pos.up();
         if (world.getBlockState(blockPos).isAir() && !world.getBlockState(blockPos).isOpaqueFullCube(world, blockPos)) {
             if (random.nextInt(100) == 0) {
-                double d = (float)pos.getX() + random.nextFloat();
-                double e = pos.getY() + 1;
-                double f = (float)pos.getZ() + random.nextFloat();
+                double d = (double)pos.getX() + random.nextDouble();
+                double e = (double)pos.getY() + 1.0;
+                double f = (double)pos.getZ() + random.nextDouble();
                 world.addParticle(ParticleTypes.LAVA, d, e, f, 0.0, 0.0, 0.0);
                 world.playSound(d, e, f, SoundEvents.BLOCK_LAVA_POP, SoundCategory.BLOCKS, 0.2f + random.nextFloat() * 0.2f, 0.9f + random.nextFloat() * 0.15f, false);
             }
