+++ net/minecraft/network/PacketByteBuf.java
@@ -352,7 +352,7 @@ extends ByteBuf {
         float g = this.readFloat();
         float h = this.readFloat();
         boolean bl = this.readBoolean();
-        return new BlockHitResult(new Vec3d((float)blockPos.getX() + f, (float)blockPos.getY() + g, (float)blockPos.getZ() + h), direction, blockPos, bl);
+        return new BlockHitResult(new Vec3d((double)blockPos.getX() + (double)f, (double)blockPos.getY() + (double)g, (double)blockPos.getZ() + (double)h), direction, blockPos, bl);
     }
 
     public void writeBlockHitResult(BlockHitResult blockHitResult) {
